Moving the HTTP actor poc to its own crate
