+++ Cargo.toml
@@ -3,6 +3,7 @@ members = [
     "crates/bin/*",
     "crates/common/*",
     "crates/core/*",
+    "crates/extensions/*",
     "crates/tests/*",
     "plugins/c8y_configuration_plugin",
     "plugins/c8y_log_plugin",
