+++ crates/extensions/tedge_http_ext/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "tedge_http_ext"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+async-trait = "0.1"
+tedge_actors = { path = "../../core/tedge_actors" }
+reqwest = { version = "0.11", default-features = false, features = ["rustls-tls"] }
